+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- feat: add plugin search and filter UI
+
 ## 1.1.0-beta.1
 
 - Feat: Refactor for WPCS and best practices